reformat disarium.cu add concept of synchronization

+++ CUDA/kernels/disarium_number.cuh
@@ -1,6 +1,6 @@
 #ifndef PRIR_DISARIUM_NUMBER_CUH
 #define PRIR_DISARIUM_NUMBER_CUH
 
-__global__ void generateDisariumNumbers(unsigned long *generatedNumbersGPU, const unsigned long NUMBERS_COUNT);
+__global__ void generateDisariumNumbers(unsigned long *generatedNumbersGPU, const unsigned int NUMBERS_COUNT);
 
 #endif //PRIR_DISARIUM_NUMBER_CUH
